translator/backend/requirements.txt
requests>=2.28.2
fastapi>=0.85.0
uvicorn>=0.19.0
python-dotenv>=0.21.0
httpx>=0.23.0
pydantic>=1.10.0